# Change some delegate FCalls to managed