--- a/src/coreclr/vm/comutilnative.h
+++ b/src/coreclr/vm/comutilnative.h
@@ -224,8 +224,9 @@ public:
 class MethodTableNative {
 public:
     static FCDECL1(UINT32, GetNumInstanceFieldBytes, MethodTable* mt);
-    static FCDECL2(FC_BOOL_RET, IsEquivalentTo, MethodTable* mta, MethodTable* mtb);
 };
+
+extern "C" BOOL QCALLTYPE MethodTable_AreTypesEquivalent(MethodTable* mta, MethodTable* mtb);
 
 class StreamNative {
 public: