--- a/src/coreclr/vm/ecalllist.h
+++ b/src/coreclr/vm/ecalllist.h
@@ -304,7 +304,6 @@ FCFuncStart(gDelegateFuncs)
     FCFuncElement("FindMethodHandle", COMDelegate::FindMethodHandle)
     FCFuncElement("AdjustTarget", COMDelegate::AdjustTarget)
     FCFuncElement("GetCallStub", COMDelegate::GetCallStub)
-    FCFuncElement("CompareUnmanagedFunctionPtrs", COMDelegate::CompareUnmanagedFunctionPtrs)
 
     // The FCall mechanism knows how to wire multiple different constructor calls into a
     // single entrypoint, without the following entry.  But we need this entry to satisfy