--- a/src/coreclr/vm/ecalllist.h
+++ b/src/coreclr/vm/ecalllist.h
@@ -603,7 +603,6 @@ FCFuncEnd()
 
 FCFuncStart(gMethodTableFuncs)
     FCFuncElement("GetNumInstanceFieldBytes", MethodTableNative::GetNumInstanceFieldBytes)
-    FCFuncElement("IsEquivalentTo", MethodTableNative::IsEquivalentTo)
 FCFuncEnd()
 
 FCFuncStart(gMngdFixedArrayMarshalerFuncs)