Benchmark:
```csharp
        [UnmanagedCallersOnly]
        static void X() { }

        private readonly Action _fnptr1 = Marshal.GetDelegateForFunctionPointer<Action>((nint)(delegate* unmanaged<void>)(&X));
        private readonly Action _fnptr2 = Marshal.GetDelegateForFunctionPointer<Action>((nint)(delegate* unmanaged<void>)(&X));

        private readonly Action _action1 = () => Console.WriteLine("A");
        private readonly Action _action2 = () => Console.WriteLine("B");

        private readonly Func<int> _actionInt1 = Func;
        private readonly Func<int> _actionInt2 = Func;

        public static int Func() => 42;

        [Benchmark]
        public bool UnmanagedEquals() => _fnptr1 == _fnptr2;

        [Benchmark]
        public Action Combine() => _action1 + _action2;

        [Benchmark]
        public bool ManagedEquals() => _actionInt1 == _actionInt2;
        
        [Benchmark]
        public bool ManagedNotEquals() => _action1 == _action2;

        [Benchmark]
        public bool TypeMismatch() => _action1.Equals(_actionInt1);
```

|           Method |        Job |         Toolchain |       Mean |     Error |    StdDev | Ratio |
|----------------- |----------- |------------------ |-----------:|----------:|----------:|------:|
|  UnmanagedEquals | Job-SQNAXT |   \PR\corerun.exe |   2.895 ns | 0.0151 ns | 0.0133 ns |  0.50 |
|  UnmanagedEquals | Job-GXCCWM | \main\corerun.exe |   5.844 ns | 0.0553 ns | 0.0490 ns |  1.00 |
|                  |            |                   |            |           |           |       |
|          Combine | Job-SQNAXT |   \PR\corerun.exe |  42.868 ns | 0.2113 ns | 0.1765 ns |  0.95 |
|          Combine | Job-GXCCWM | \main\corerun.exe |  45.173 ns | 0.3683 ns | 0.3445 ns |  1.00 |
|                  |            |                   |            |           |           |       |
|    ManagedEquals | Job-SQNAXT |   \PR\corerun.exe |   6.934 ns | 0.0747 ns | 0.0624 ns |  0.58 |
|    ManagedEquals | Job-GXCCWM | \main\corerun.exe |  11.938 ns | 0.0670 ns | 0.0560 ns |  1.00 |
|                  |            |                   |            |           |           |       |
| ManagedNotEquals | Job-SQNAXT |   \PR\corerun.exe |  94.570 ns | 0.7560 ns | 0.7072 ns |  0.94 |
| ManagedNotEquals | Job-GXCCWM | \main\corerun.exe | 100.963 ns | 1.5713 ns | 1.4697 ns |  1.00 |
|                  |            |                   |            |           |           |       |
|     TypeMismatch | Job-SQNAXT |   \PR\corerun.exe |   1.587 ns | 0.0117 ns | 0.0097 ns |  0.17 |
|     TypeMismatch | Job-GXCCWM | \main\corerun.exe |   9.482 ns | 0.0368 ns | 0.0326 ns |  1.00 |